Merge remote-tracking branch 'origin/4.5.1'

+++ sql/log_event.cc
@@ -2752,8 +2752,9 @@ Query_log_event::Query_log_event(const char* buf, uint event_len,
   
   slave_proxy_id= thread_id = uint4korr(buf + Q_THREAD_ID_OFFSET);
   exec_time = uint4korr(buf + Q_EXEC_TIME_OFFSET);
-  db_len = (uint)buf[Q_DB_LEN_OFFSET]; // TODO: add a check of all *_len vars
-  error_code = uint2korr(buf + Q_ERR_CODE_OFFSET);
+  // @InfiniDB @bug5910 fix for db_len=128
+  db_len = (uchar)buf[Q_DB_LEN_OFFSET]; // TODO: add a check of all *_len vars
+   error_code = uint2korr(buf + Q_ERR_CODE_OFFSET);
 
   /*
     5.0 format starts here.
